requirements.txt
autopep8==1.4.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.2
django-allauth==0.44.0
django-crispy-forms==1.11.0
idna==2.10
oauthlib==3.1.0
pep8==1.7.1
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2018.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.3
